% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_spec.R
\name{arma_spec}
\alias{arma_spec}
\title{Calculates the spectral density for a known ARMA process}
\usage{
arma_spec(omega, phi = 0, theta = 0)
}
\arguments{
\item{omega}{A vector containing a set of frequencies the spectral density is to be calculated over.}

\item{phi}{The vector of AR coefficient(s).}

\item{theta}{The vector of MA coefficient(s).}
}
\value{
A vector as long as the \code{omega} vector containing the calculated spectral density
}
\description{
\code{arma_spec()} gives the spectral density across a given set of \eqn{\omega}s for a known ARMA process.
This is an extra function to help users compare spectral density estimates obtained from
\code{\link[=HBEST]{HBEST()}} or \code{\link[=HBEST_fast]{HBEST_fast()}} against the "truth" when the data generation process is known.
}
\examples{
omega <- seq(0, pi, length.out = 1000)
phi <- 0.5
arma_spec(omega = omega, phi = phi)
## output is a vector of length 1000

}
