% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_spec_vectorized.R
\name{arma_spec_vectorized}
\alias{arma_spec_vectorized}
\title{\code{arma_spec_vectorized}: A vectorized version of \code{\link[=arma_spec]{arma_spec()}}}
\usage{
arma_spec_vectorized(omega, phi = 0, theta = 0)
}
\arguments{
\item{omega}{A vector containing a set of frequencies the spectral density is to be calculated over.}

\item{phi}{The vector of AR coefficient(s).}

\item{theta}{The vector of MA coefficient(s).}
}
\value{
A vector as long as the \code{omega} vector containing the calculated spectral density.
}
\description{
Gives the spectral density across a given set of \eqn{\omega}s for a known ARMA process.
This function is vectorized for faster computation.
}
\examples{
omega <- seq(0, pi, length.out = 1000)
phi <- 0.5
arma_spec_vectorized(omega = omega, phi = phi)
## output is a vector of length 1000

}
