% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol_sampling.R
\name{chol_sampling}
\alias{chol_sampling}
\title{Perform Cholesky Sampling}
\usage{
chol_sampling(Lt, d, beta_c)
}
\arguments{
\item{Lt}{A \verb{(d x d)} Upper triangular matrix from Cholesky decomposition}

\item{d}{A numeric strictly positive value denoting the length of beta}

\item{beta_c}{A vector of length \code{d} of center beta values}
}
\value{
a vector of proposed beta values
}
\description{
This function performs Cholesky Sampling given the Cholesky decomposition and the centers of the beta coefficients.
}
