% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Function to call different data generation methods for simulations.}
\usage{
gen_data(gen_method, ...)
}
\arguments{
\item{gen_method}{}

\item{...}{}
}
\description{
Function to call different data generation methods for simulations.
}
