% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_AR.R
\name{generate_AR}
\alias{generate_AR}
\title{Generate R time series from the same AR(p) process}
\usage{
generate_AR(phi, n = 1000, R = 1, burn = 50)
}
\arguments{
\item{phi}{A vector that contains the AR(p) coefficients. Same syntax used in \code{arima.sim()}, e.g. \code{c(ar1coefficient, ar2coefficient)}.}

\item{n}{An optional numeric scalar (default \code{1000}) that determines the length of the time series generated.}

\item{R}{An optional numeric scalar (default \code{1}) that determines how many time series to generate. Each time series is stored column wise in output.}

\item{burn}{An optional numeric scalar (default \code{50}) that determines the \code{n.start} argument of the \code{arima.sim()} function.}
}
\value{
a list object that contains the matrix object \code{matrix_timeseries} of dimension \verb{(n x R)}
}
\description{
\code{generate_AR} creates \code{R} copies of an AR(p) process from \code{arima.sim()}.
}
\examples{
## An AR(1) process with `\phi = 0.5`
phi <- c(0.5)
n <- 2000
R <- 5
burn <- 100
generate_AR(phi = c(0.5), n = n, R = R, burn = 100)
## output generates a list that contains a (2000 x 5) matrix

}
