% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_AR1_vary.R
\name{generate_AR1_vary}
\alias{generate_AR1_vary}
\title{Generate multiple AR(1) varying time series}
\usage{
generate_AR1_vary(n, R = 1, min = 0.45, max = 0.6, burn = 50)
}
\arguments{
\item{n}{An optional numeric scalar (default \code{1000}) that determines the length of the time series generated.}

\item{R}{An optional numeric scalar (default \code{1}) that determines how many time series to generate. Each time series is stored column wise in output.}

\item{min}{An optional numeric scalar (default \code{0.45}) that determines the minimum value the AR(1) coefficient could possibly take.}

\item{max}{An optional numeric scalar (default \code{0.60}) that determines the maximum value the AR(1) coefficient could possibly take.}

\item{burn}{An optional numeric scalar (default \code{50}) that determines the \code{n.start} argument of the \code{arima.sim()} function.}
}
\value{
a list object that contains the following fields:
\itemize{
\item \code{matrix_timeseries}: a \verb{(n x R)} matrix that contains the slightly varying AR(1) time series.
\item \code{phi_true}: a \code{R} long vector object that contains the AR(1) coefficients for each time series.
}
}
\description{
Generates \code{R} time series from an AR(1) varying processes.
Each time series AR(1) coefficient is uniformly chosen from the range provided by the \code{min} and \code{max} arguments.
}
\examples{
n <- 1000
R <- 5
min <- 0.45
max <- 0.6
burn <- 50
generate_AR1_vary(n = 1000, R = 5, min = 0.45, max = 0.6, burn = 50)
## Output is a list containing a (1000 x 5) matrix and a 5 long vector containing the true phi values randomly generated to create the R many time series.

}
