% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MA1.R
\name{generate_MA1}
\alias{generate_MA1}
\title{Generate R many time series from a random MA(1) process}
\usage{
generate_MA1(n = 1000, R = 1, burn = 50)
}
\arguments{
\item{n}{An optional numeric scalar (default \code{1000}) that determines the length of the time series generated.}

\item{R}{An optional numeric scalar (default \code{1}) that determines how many time series to generate. Each time series is stored column wise in output.}

\item{burn}{An optional numeric scalar (default \code{50}) that determines the \code{n.start} argument of the \code{arima.sim()} function.}
}
\value{
a list object that contains the following fields:
\itemize{
\item \code{matrix_timeseries}: a \verb{(n x R)} matrix that contains the slightly varying MA(1) time series.
\item \code{theta_true}: a \code{R} long vector object that contains the MA(1) coefficients generated for each time series.
}
}
\description{
Generate R many time series from an MA(1) process with a randomly chosen MA(1) coefficient
from a Uniform(0,1) distribution.
}
\examples{
n <- 2000
R <- 20
burn <- 50
generate_MA1(n = 2000, R = 20, burn = 50)
## Returns a list with a matrix (2000 x 20) and
## a R long vector of generated theta values

}
