% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MA4_vary.R
\name{generate_MA4_vary}
\alias{generate_MA4_vary}
\title{Title}
\usage{
generate_MA4_vary(n = 1000, R = 1, burn = 50, alpha = 0.05)
}
\arguments{
\item{n}{}

\item{R}{}

\item{burn}{}

\item{alpha}{}
}
\description{
Title
}
