% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MA4_varyType2.R
\name{generate_MA4_varyType2}
\alias{generate_MA4_varyType2}
\title{Title}
\usage{
generate_MA4_varyType2(n, R = 1, burn = 50, alpha)
}
\arguments{
\item{n}{}

\item{R}{}

\item{burn}{}

\item{alpha}{}
}
\description{
Title
}
