% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_glob_HBEST.R
\name{grad_glob_HBEST}
\alias{grad_glob_HBEST}
\title{Calculate the gradient of the conditional log posterior distribution of \eqn{\pmb{a}} from HBEST.}
\usage{
grad_glob_HBEST(ebr, ab, sumPsi, Psi_list, y_list, Sigma_a, R)
}
\arguments{
\item{ebr}{}

\item{ab}{}

\item{sumPsi}{((B+1) x R) The \code{r}th column of the column sum across \code{Psi}.}

\item{Psi_list}{}

\item{y_list}{}

\item{Sigma_a}{}

\item{R}{}
}
\value{
A vector containing the gradient for the conditional posterior of \eqn{\beta}.
}
\description{
\code{grad_glob_HBEST} function calculates the conditional posterior for the \code{r}th \eqn{\beta} coefficients under HBEST.
}
