% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_loc_HBEST.R
\name{grad_loc_HBEST}
\alias{grad_loc_HBEST}
\title{Calculate the gradient of the conditional log posterior distribution of the \eqn{\pmb{e}^{(r)}} from Model A.}
\usage{
grad_loc_HBEST(er, ab, sumPsi, Psi, y, Sigma_e)
}
\arguments{
\item{er}{}

\item{ab}{}

\item{sumPsi}{The \code{r}th column of the column sum across \code{Psi}.}

\item{Psi}{The \code{r}th \code{Psi} matrix.}

\item{y}{The \code{r}th periodogram sub-setted for the Fourier frequencies for the \code{r}th time series.}

\item{Sigma_e}{}
}
\value{
A vector containing the gradient for the conditional posterior of \eqn{\beta}.
}
\description{
\code{grad_loc_HBEST} function calculates the conditional posterior for the \code{r}th \eqn{\beta} coefficients under Model B.
}
