% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradienta_modelA_fast.R
\name{gradienta_modelA_fast}
\alias{gradienta_modelA_fast}
\title{Calculate the gradient of the conditional log posterior distribution of \eqn{\pmb{a}} from Model A.}
\usage{
gradienta_modelA_fast(
  ab,
  sumPsi,
  Psi_list,
  y_list,
  Sigma_a,
  R,
  sumsumPsi,
  Psi_ebr_list
)
}
\arguments{
\item{ab}{}

\item{sumPsi}{((B+1) x R) The \code{r}th column of the column sum across \code{Psi}.}

\item{Psi_list}{}

\item{y_list}{}

\item{Sigma_a}{}

\item{R}{}

\item{ebr}{}
}
\value{
A vector containing the gradient for the conditional posterior of \eqn{\beta}.
}
\description{
\code{gradienta_modelA_fast} function calculates the conditional posterior for the \code{r}th \eqn{\beta} coefficients under Model B.
}
