% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradiente_modelA_fast.R
\name{gradiente_modelA_fast}
\alias{gradiente_modelA_fast}
\title{Calculate the gradient of the conditional log posterior distribution of the \eqn{\pmb{e}^{(r)}} from Model A.}
\usage{
gradiente_modelA_fast(er, sumPsi, Psi, y, Sigma_e, Psi_ab)
}
\arguments{
\item{er}{}

\item{sumPsi}{The \code{r}th column of the column sum across \code{Psi}.}

\item{Psi}{The \code{r}th \code{Psi} matrix.}

\item{y}{The \code{r}th periodogram sub-setted for the Fourier frequencies for the \code{r}th time series.}

\item{Sigma_e}{}

\item{ab}{}
}
\value{
A vector containing the gradient for the conditional posterior of \eqn{\beta}.
}
\description{
\code{gradiente_modelA_fast} function calculates the conditional posterior for the \code{r}th \eqn{\beta} coefficients under Model B.
}
