% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hea_modelA_fast.R
\name{hea_modelA_fast}
\alias{hea_modelA_fast}
\title{Hessian function for conditional log posterior of \eqn{\pmb{a}} from Model A}
\usage{
hea_modelA_fast(ab, Psi_list, y_list, ebr, Sigma_a, R, Psi_ebr_list)
}
\arguments{
\item{ab}{}

\item{Psi_list}{}

\item{y_list}{}

\item{ebr}{}

\item{R}{}

\item{Sigma_e}{}
}
\description{
\code{hea_modelA} calculates the hessian of the conditional posterior of the \code{r}th \eqn{beta} for model A.
}
