% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hee_modelA_fast.R
\name{hee_modelA_fast}
\alias{hee_modelA_fast}
\title{Hessian function for conditional log posterior of \eqn{\pmb{e}^{(r)}}  from Model A}
\usage{
hee_modelA_fast(Psi, y, er, ab, Sigma_e, Psi_ab)
}
\arguments{
\item{zetasquared_r}{}
}
\description{
\code{hee_modelA_fast} calculates the hessian of the conditional posterior of the \code{r}th \eqn{beta} for model A.
}
