% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hess_glob_HBEST.R
\name{hess_glob_HBEST}
\alias{hess_glob_HBEST}
\title{Hessian function for conditional log posterior of \eqn{\pmb{a}} from Model A}
\usage{
hess_glob_HBEST(glob, Psi_list, y_list, loc, Sigma_glob, R)
}
\arguments{
\item{glob}{}

\item{Psi_list}{}

\item{y_list}{}

\item{loc}{}

\item{Sigma_glob}{}

\item{R}{}
}
\description{
\code{hess_glob_HBEST} calculates the hessian of the conditional posterior of the \code{r}th \eqn{beta} for model A.
}
