% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hess_loc_HBEST.R
\name{hess_loc_HBEST}
\alias{hess_loc_HBEST}
\title{Hessian function for conditional log posterior of \eqn{\pmb{beta}_{(r)}}  from HBESt}
\usage{
hess_loc_HBEST(Psi, y, loc, glob, Sigma_loc)
}
\arguments{
\item{y}{}

\item{loc}{}

\item{glob}{}

\item{Sigma_loc}{}
}
\description{
\code{hess_loc_HBEST} calculates the hessian of the conditional posterior of the \code{r}th \eqn{beta} for HBEST.
}
