% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpost_glob_HBEST.R
\name{logpost_glob_HBEST}
\alias{logpost_glob_HBEST}
\title{Function that calculates the conditional log posterior distribution of the vector \eqn{\pmb{a}} for Model A.}
\usage{
logpost_glob_HBEST(glob, loc, Sigma_glob, y_list, sumPsi, Psi_list, R)
}
\arguments{
\item{glob}{}

\item{loc}{}

\item{Sigma_glob}{}

\item{y_list}{}

\item{sumPsi}{}

\item{Psi_list}{}

\item{R}{}
}
\value{
A vector that contains the \code{r}th posterior conditional for \eqn{\beta_{(r)}}.
}
\description{
\code{logpost_glob_HBEST} calculates the conditional posterior distribution for model B
}
