% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpost_loc_HBEST.R
\name{logpost_loc_HBEST}
\alias{logpost_loc_HBEST}
\title{Function that calculates the conditional log posterior distribution of \eqn{\pmb{\beta}_{(r)}} from HBEST.}
\usage{
logpost_loc_HBEST(sumPsi, loc, glob, Psi, y, Sigma_loc)
}
\arguments{
\item{sumPsi}{}

\item{loc}{}

\item{glob}{}

\item{Psi}{}

\item{y}{}

\item{Sigma_loc}{}
}
\value{
A vector that contains the \code{r}th posterior conditional for \eqn{\beta^{(r)}}.
}
\description{
\code{logpost_loc_HBEST} calculates the conditional posterior distribution for HBEST
}
