% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logposteriora_modelA_fast.R
\name{logposteriora_modelA_fast}
\alias{logposteriora_modelA_fast}
\title{Function that calculates the conditional log posterior distribution of the vector \eqn{\pmb{a}} for Model A.}
\usage{
logposteriora_modelA_fast(
  ab,
  Sigma_a,
  y_list,
  sumPsi,
  Psi_list,
  R,
  sumsumPsi,
  Psi_ebr_list
)
}
\arguments{
\item{ab}{}

\item{Sigma_a}{}

\item{y_list}{}

\item{sumPsi}{}

\item{Psi_list}{}

\item{ebr}{}
}
\value{
A vector that contains the \code{r}th posterior conditional for \eqn{\beta^{(r)}}.
}
\description{
\code{logposteriora_modelA} calculates the conditional posterior distribution for model B
}
