% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logposteriore_modelA_fast.R
\name{logposteriore_modelA_fast}
\alias{logposteriore_modelA_fast}
\title{Function that calculates the conditional log posterior distribution of \eqn{\pmb{e}^{(r)}} from Model A.}
\usage{
logposteriore_modelA_fast(sumPsi, er, Psi, y, Sigma_e, Psi_ab)
}
\arguments{
\item{sumPsi}{}

\item{Psi}{}

\item{y}{}

\item{br}{}

\item{Sigma}{}

\item{zetasquared_r}{}
}
\value{
A vector that contains the \code{r}th posterior conditional for \eqn{\beta^{(r)}}.
}
\description{
\code{logposteriore_modelA} calculates the conditional posterior distribution for model B
}
