% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler_Common.R
\name{Sampler_Common}
\alias{Sampler_Common}
\title{Sampler to run the Common method}
\usage{
Sampler_Common(
  ts_list,
  B,
  iter,
  nu_tau,
  sigmasquared,
  tausquared,
  tau_min,
  tau_max,
  num_gpts,
  burnin
)
}
\arguments{
\item{ts_list}{}

\item{B}{}

\item{iter}{}

\item{tausquared}{}

\item{nu}{}

\item{etasq}{}

\item{sigmasquare}{}

\item{lambda}{}
}
\description{
Sampler to run the Common method
}
