% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tausquared_common.R
\name{tausquared_common}
\alias{tausquared_common}
\title{Griddy Gibbs Update for Common and Independent Method.}
\usage{
tausquared_common(b, B, D, nu_tau, tau_min, tau_max, num_gpts)
}
\arguments{
\item{b}{:}

\item{B}{: The number of basis functions.}

\item{D}{:}

\item{nu_tau}{: Degrees of freedom used for the t-distribution CDF and quantile functions.}

\item{tau_min}{: The minimum value for the creation of the uniform grid of \eqn{\tau} values.}

\item{tau_max}{: The maximum value for the creation of the uniform grid of \eqn{\tau} values.}

\item{num_gpts}{: number of grid points controls how dense the grid is.}
}
\description{
Griddy Gibbs Update for Common and Independent Method.
}
