% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tausquared_HBEST.R
\name{tausquared_HBEST}
\alias{tausquared_HBEST}
\title{Griddy Gibbs update for global smoothing parameter (\code{tausquared})}
\usage{
tausquared_HBEST(
  loc,
  glob,
  B,
  D,
  R,
  cur_zetasquared,
  nu_tau,
  tau_min,
  tau_max,
  num_gpts
)
}
\arguments{
\item{loc}{}

\item{glob}{}

\item{B}{}

\item{D}{}

\item{R}{}

\item{cur_zetasquared}{}

\item{nu_tau}{}

\item{tau_min}{}

\item{tau_max}{}

\item{num_gpts}{}
}
\description{
A function which samples values for the \code{tausquared} update as specified in HBEST.
}
