% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittle_post.R
\name{whittle_post}
\alias{whittle_post}
\title{Conditional Posterior of Beta coefficients}
\usage{
whittle_post(ab, X, sumX, tsq, perio, sigmasalpha, D)
}
\arguments{
\item{ab}{A \code{(B+1)} column vector of intercept and basis function coefficient values.}

\item{X}{A basis vector matrix.}

\item{sumX}{A vector of the column sums of the \code{X} matrix.}

\item{tsq}{A tau squared scalar.}

\item{perio}{A vector of the log of the periodogram.}

\item{sigmasalpha}{A prior intercept variance, variance associated with the alpha_0 prior}

\item{D}{A vector of length B with strictly positive entries.}
}
\value{
A scalar of the value of the conditional posterior at the given value
}
\description{
Conditional Posterior of Beta coefficients
}
