% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zetasquared_HBEST.R
\name{zetasquared_HBEST}
\alias{zetasquared_HBEST}
\title{Griddy Gibbs update for the local parameter zetasquared}
\usage{
zetasquared_HBEST(loc, B, D, tausquared, nu_zeta, zeta_min, zeta_max, num_gpts)
}
\arguments{
\item{B}{An integer specifying the number of basis coefficients.}

\item{D}{The vector of dampening coefficients in the prior.}

\item{tausquared}{The current global smoothing parameter \code{tausquared}.}

\item{nu_zeta}{A scalar that controls the degrees of freedom for the \code{zeta} prior.}

\item{zeta_min}{The smallest value zeta can take (default to \code{1}).}

\item{zeta_max}{The largest value zeta can take (default to \code{10}).}

\item{num_gpts}{A number which controls how dense the grid is.}
}
\value{
A vector containing the sampled zeta values.
}
\description{
only need to pass beta_cur_sq_norm = (\eqn{\beta' D^{-1} \beta}). \eqn{\zeta} has half-t prior (with df = \code{nu_eta})
}
