% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Function to call different data generation methods for simulations.}
\usage{
gen_data(gen_method, ...)
}
\arguments{
\item{gen_method}{Which method of data generation to use. Options include \code{ARp}, \code{AR1vary}, \code{AR2mix}, \code{MA4}, \code{MA4vary}, and \code{MA4varyType2}.}

\item{...}{}
}
\description{
Function to call different data generation methods for simulations.
}
