% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Time-series Data Generation Wrapper Function.}
\usage{
gen_data(gen_method, ...)
}
\arguments{
\item{gen_method}{Method of data generation to use. Options include \code{ARp}, \code{AR1vary}, \code{AR2mix}, \code{MA4}, \code{MA4vary}, and \code{MA4varyType2}.}

\item{...}{further arguments, passed to chosen method. See details below.}
}
\value{
The \code{gen_data()} function returns the appropriate objects based on the method chosen. See help pages for each function for more details.
}
\description{
This function is a wrapper function for various time-series data generation methods.
This is used to call one of six methods that generates data in the structure used for simulations for \code{\link[=HBEST]{HBEST()}} or \code{\link[=HBEST_fast]{HBEST_fast()}}.
}
\details{
Available data generation methods include:
\tabular{ll}{
\code{ARp} \tab calls the \code{\link[=generate_AR]{generate_AR()}} function which is based on \link[stats:arima.sim]{stats::arima.sim}. \cr
\code{AR1vary} \tab calls the \code{\link[=generate_AR1_vary]{generate_AR1_vary()}} function which is based on \link[stats:arima.sim]{stats::arima.sim}. \cr
\code{AR2mix} \tab calls the \code{\link[=generate_AR2_mixture]{generate_AR2_mixture()}} function which is based on \link[stats:arima.sim]{stats::arima.sim} and \insertCite{granados-garcia_brain_2022}{HBEST}. \cr
\code{MA4} \tab calls the \code{\link[=generate_MA4]{generate_MA4()}} function which is based on \link[stats:arima.sim]{stats::arima.sim}. \cr
\code{MA4vary} \tab calls the \code{\link[=generate_MA4_vary]{generate_MA4_vary()}} function which is based on \link[stats:arima.sim]{stats::arima.sim}. \cr
\code{MA4varyType2} \tab calls the \code{\link[=generate_MA4_varyType2]{generate_MA4_varyType2()}} function which is based on \link[stats:arima.sim]{stats::arima.sim}. \cr
}
}
\examples{
R = 20
n = 500
burn = 50

## `ARp` method:
ts = gen_data(gen_method = "ARp", phi = c(0.5), n = n, R = R, burn = burn)

## `AR1vary` method:
ts = gen_data(gen_method = "AR1vary", n = n, R = 5, min = 0.45, max = 0.6, burn = burn)

## `AR2mix` method:
R = 20
peaks1 = runif(R, min = 0.2, max = 0.23)
bandwidths1 = runif(R, min = .1, max = .2)
peaks2 = runif(R, min = (pi * (2/5)) - 0.1, max = (pi * (2/5)) + 0.1)
bandwidths2 = rep(0.15, R)
peaks = rbind(peaks1, peaks2)
bandwidths = rbind(bandwidths1, bandwidths2)
ts = gen_data(gen_method = "AR2mix", peaks = peaks, bandwidths = bandwidths, n_vary = c(rep(300, 10), rep(800, 10)))

## `MA4` method:
n = rep(500, R)
ts = gen_data(gen_method = "MA4", n = n, R = R, burn = burn)

## `MA4vary` method:
n = rep(500, R)
alpha = 0.05
ts = gen_data(gen_method = "MA4vary", n = n, R = R, burn = burn, alpha = alpha)

## `MA4varyType2` method:
n = rep(500, R)
ts = gen_data(gen_method = "MA4varyType2", n = n, R = R, burn = burn, alpha = alpha)

}
