% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_AR.R
\name{generate_AR}
\alias{generate_AR}
\title{Generate \code{R}-Many Conditionally Independent Time Series from the Same AR(p) Process}
\usage{
generate_AR(phi, n = 1000, R = 1, burn = 50)
}
\arguments{
\item{phi}{A vector containing the AR(p) coefficients. Same syntax used in \code{arima.sim()}, e.g. \code{c(ar1coefficient, ar2coefficient)}.}

\item{n}{An optional numeric scalar (default \code{1000}) that determines the length of the time series generated.}

\item{R}{An optional numeric scalar (default \code{1}) that determines how many time series to generate. Each time series is stored column wise in output.}

\item{burn}{An optional numeric scalar (default \code{50}) that determines the \code{n.start} argument of the \code{arima.sim()} function.}
}
\value{
a list object that contains the \code{R}-long list object \code{ts_list} each containing a matrix of dimension \verb{(n x 1)} and the \code{true_phi} which is the user-specified \code{phi}.
}
\description{
\code{generate_AR} generates \code{R}-many conditionally independent time series from an AR(p) process from \code{arima.sim()}.
}
\examples{
## An AR(1) process with `\phi = 0.5`
phi <- c(0.5)
n <- 2000
R <- 5
burn <- 100
ts <- generate_AR(phi = c(0.5), n = n, R = R, burn = 100)$ts_list
## output generates an R-long list that each contains a (2000 x 1) matrix

}
