% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_AR2_mixture.R
\name{generate_AR2_mixture}
\alias{generate_AR2_mixture}
\title{Generate \code{R}-Many Conditionally Independent Varying AR(2) Time Series with Angle Representation}
\usage{
generate_AR2_mixture(peaks, bandwidths, n_vary, variances = NULL)
}
\arguments{
\item{peaks}{a vector of length \code{R} with peak locations of each spectrum.}

\item{bandwidths}{a vector of length \code{n} with the bandwidths of each spectrum.}

\item{n_vary}{a numeric strictly positive scalar that determines the length of the time series generated.}

\item{variances}{an optional vector (default is \code{NULL}) of length \code{n} with the variances of the innovations.}
}
\value{
a list object that contains the following fields:
\itemize{
\item \code{matrix_timeseries}: a \verb{(n x R)} matrix that contains the slightly varying AR(2) time series.
\item \code{phi}: a \verb{(2 x R)} matrix object that contains the AR(2) coefficients generated for each time series.
}
}
\description{
Generates \code{R} many slightly varying AR(2) time series given a \code{R} long vector of peak locations and and \code{R} long vector of bandwidths
}
\examples{
R = 20
peaks1 = runif(R, min = 0.2, max = 0.23)
bandwidths1 = runif(R, min = .1, max = .2)
peaks2 = runif(R, min = (pi * (2/5)) - 0.1, max = (pi * (2/5)) + 0.1)
bandwidths2 = rep(0.15, R)
peaks = rbind(peaks1, peaks2)
bandwidths = rbind(bandwidths1, bandwidths2)
ts = generate_AR2_mixture(peaks = peaks, bandwidths = bandwidths, n_vary = c(rep(300, 10), rep(800, 10)))
## Returns an R-long list object each with a (n_vary[r] x 1) matrix object,
## and a (2 x 20) matrix of AR(2) coefficients

}
