% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MA4_vary.R
\name{generate_MA4_vary}
\alias{generate_MA4_vary}
\title{generate \code{R}-many conditionally independent varying MA(4) time series}
\usage{
generate_MA4_vary(n, R = 1, burn = 50, alpha = 0.05)
}
\arguments{
\item{n}{A numeric vector that determines the length of the time series generated. Must contain \code{R}-many entries. Time series may be of different lengths.}

\item{R}{An optional scalar indicating the number of conditionally independent time series to be generated. (default is \code{1}).}

\item{burn}{A scalar indicating the amount of burn-in to be used with \code{\link[stats:arima.sim]{stats::arima.sim()}}.  (default is \code{50}).}

\item{alpha}{A scalar specifying the variation wanted from the base coefficients. (default is \code{0.05}).}
}
\value{
The function returns a list containing:
\tabular{ll}{
\code{ts_list} \tab returns an \code{R}-long list each containing an \verb{(n[r]} \eqn{\times} \verb{1)} matrix of the generated time series. \cr
\code{theta_true} \tab returns a \verb{(4} \eqn{\times} \verb{R)} matrix of true generated MA(4) coefficients. \cr
\code{alpha} \tab returns the user-provided variation scalar. \cr
\code{mu_r_gen} \tab returns the \verb{(4} \eqn{\times} \verb{R)} matrix of the generated standard normal values used to help calculate the new MA(4) coefficients. \cr
}
}
\description{
\code{generate_MA4_vary()} generates \code{R}-many time series where the MA(4) that is used as the base coefficients for the variation is \verb{(-.3,-.6,-.3,.6)}.
This is based on the MA(4) used in \insertCite{granados-garcia_brain_2022;textual}{HBEST}.
}
\details{
The variation around the base theta coefficients is generated in the function by:
\enumerate{
\item Sample 4 values from a \eqn{N(0,1)} to use as a "mean".
\item Calculate: \code{basetheta + alpha * mu_r * abs(basetheta)} where \code{basetheta} is the original coefficients; \code{mu_r} is the 4 sampled values from the \eqn{N(0,1)}; and \code{alpha} is the user-specified scalar that controls the variation around the base coefficients.
\item Generate: A new time-series using the new \code{theta} value from step 2.
\item Repeat steps 1-3 \code{R}-many times.
}
}
\examples{
R <- 20
## For time series of different lengths:
n <- c(rep(500, R/2), rep(800, R/2))
burn <- 50
alpha <- 0.05
ts <- generate_MA4_vary(n = n, R = R, burn = burn, alpha = alpha)$ts_list

## The function returns an R-long list object each with a (n[r] x 1) matrix object,
## a (4 x 20) matrix of true MA(4) coefficients,
## a scalar returning the alpha provided,
## and a (4 x 20) matrix of the standard normal values generated for each R.

## Plot
## Create an empty plot
plot(
  x = c(),
  y = c(),
  xlim = c(0, 800),
  ylim = range(ts),
  ylab = "",
  xlab = "time"
)
for (r in 1:10) {
  lines(ts[[r]][, 1], col = "blue")
}
for (r in 11:R) {
  lines(ts[[r]][, 1], col = "red")
}

}
