% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MA4_vary.R
\name{generate_MA4_vary}
\alias{generate_MA4_vary}
\title{Generate Conditionally Independent \code{R}-Many MA(4) Varying Time Series of the Same Length}
\usage{
generate_MA4_vary(n = 1000, R = 1, burn = 50, alpha = 0.05)
}
\arguments{
\item{n}{A scalar indicating the length of all \code{R} time series.}

\item{R}{A scalar indicating the number of conditionally independent time series to be generated.}

\item{burn}{A scalar indicating the amount of burn-in to be used with \code{\link[stats:arima.sim]{stats::arima.sim()}}.}

\item{alpha}{A scalar indicating the}
}
\description{
The MA(4) that is used as the base coefficients for the variation is \verb{(-.3,-.6,-.3,.6)}. This is based on the MA(4) used in \insertRef{granados-garcia_brain_2022}{HBEST}.
}
\details{
The variation around the base theta coefficients is generated by:
\enumerate{
\item Sample 4 values from a \eqn{N(0,1)} to use as a "mean".
\item Calculate: \code{basetheta + alpha * mu_r * abs(basetheta)} where \code{basetheta} is the original coefficients; \code{mu_r} is the 4 sampled values from the \eqn{N(0,1)}; and \code{alpha} is the user-specified scalar that controls the variation around the base coefficients.
\item Generate: A new time-series using the new \code{theta} value from step 2.
}
}
