% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MA4_varyType2.R
\name{generate_MA4_varyType2}
\alias{generate_MA4_varyType2}
\title{Generate Conditionally Independent \code{R}-Many MA(4) Varying Time Series of the Same Length (Version 2)}
\usage{
generate_MA4_varyType2(n, R = 1, burn = 50, alpha)
}
\arguments{
\item{n}{A scalar indicating the length of all \code{R} time series. (default is \code{1000}).}

\item{R}{A scalar indicating the number of conditionally independent time series to be generated. (default is \code{1}).}

\item{burn}{A scalar indicating the amount of burn-in to be used with \code{\link[stats:arima.sim]{stats::arima.sim()}}.  (default is \code{50}).}

\item{alpha}{A scalar specifying the variation wanted from the first base coefficient. (default is \code{0.05}).}
}
\value{
The function returns a list containing:
\tabular{ll}{
\code{ts_list} \tab returns an \code{R}-long list each containing an \verb{(n} \eqn{\times} \verb{1)} matrix of the generated time series. \cr
\code{theta_true} \tab returns a \verb{(4} \eqn{\times} \verb{R)} matrix of true generated MA(4) coefficients. \cr
}
}
\description{
The MA(4) that is used as the base coefficients for the variation is \verb{(-.3,-.6,-.3,.6)}. In this version of the data generation, only the first coefficient is changed, all other coefficients are left as is.
This is based on the MA(4) used in \insertCite{granados-garcia_brain_2022;textual}{HBEST}.
This is the data generation that is used in \insertCite{lee_hierarchical_2025;textual}{HBEST}.
}
\details{
The variation around the first base theta coefficient is generated by:
\enumerate{
\item Sample 1 value from a \eqn{N(0,1)} to use as a "mean".
\item Calculate: \code{basetheta[1] + alpha * mu_r * abs(basetheta[1])} where \code{basetheta[1]} is the original first MA(4) coefficient; \code{mu_r} is the 1 sampled values from the \eqn{N(0,1)}; and \code{alpha} is the user-specified scalar that controls the variation around the first base coefficient.
\item Generate: A new time-series using the new \code{theta} value from step 2.
\item Repeat steps 1-3 \code{R}-many times.
}
}
\examples{
R = 20
n = 500
burn = 50
alpha = 0.05
ts = generate_MA4_varyType2(n = n, R = R, burn = burn, alpha = alpha)

# plot the time series generated:
# create an empty plot
plot(x = c(), y = c(), xlim = c(0,n), ylim = range(ts$ts_list), main = "Example", ylab = "", xlab = "time")
for(r in 1:R){
lines(ts$ts_list[[r]][,1])
}
## Returns an R-long list object each with a (500 x 1) matrix object,
## a (4 x 20) matrix of true MA(4) coefficients,
## a scalar returning the alpha provided,
## and a (4 x 20) matrix of the standard normal values generated for each R.

}
