% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler_HBEST_fast.R
\name{HBEST_fast}
\alias{HBEST_fast}
\title{A Sampling Algorithm for HBEST (fast version)}
\usage{
HBEST_fast(
  ts_list,
  B,
  iter,
  sigmasquared_a,
  sigmasquared_e,
  nu_tau,
  tausquared,
  nu_zeta,
  burnin,
  zeta_min,
  zeta_max,
  tau_min,
  tau_max,
  num_gpts
)
}
\arguments{
\item{ts_list}{A list \code{R} long containing the vectors of the stationary time series of potentially different lengths.}

\item{B}{An integer specifying the number of basis coefficients (not including the intercept basis coefficient \eqn{\beta_0})}

\item{iter}{An integer specifying the number of iterations for the MCMC algorithm embedded in this function.}

\item{nu_tau}{A scalar indicating the degrees of freedom for the prior on \eqn{\tau}.}

\item{tausquared}{A scalar which is used as the initial value of \code{tausquared} that controls the global smoothing effect.}

\item{nu_zeta}{A scalar indicating the degrees of freedom for the prior on \eqn{\zeta}.}

\item{burnin}{An integer specifying the burn-in to be removed at the end of the sampling algorithm.}

\item{zeta_min}{A scalar controlling the smallest value \eqn{\zeta} can take. So, \code{zeta_min}^2 is the smallest value \code{zetasquared} can take.}

\item{zeta_max}{A scalar controlling the largest value \eqn{\zeta} can take. So, \code{zeta_max}^2 is the largest value that \code{zetasquared} can take.}

\item{tau_min}{A scalar controlling the smallest value \eqn{\tau} can take. So, \code{tau_min}^2 is the smallest value \code{tausquared} can take.}

\item{tau_max}{A scalar controlling the largest value \eqn{\tau} can take. So, \code{tau_max}^2 is the largest value \code{tausquared} can take.}

\item{num_gpts}{A scalar controlling the denseness of the grid during the sampling of both \code{tausquared} and \code{zetasquared}.}

\item{sigmasquared}{A scalar specifying the variance of the prior (N(0,\code{sigmasquared})) intercept term for \eqn{\beta_0}. (default is \code{100} to ensure a diffuse prior).}
}
\description{
\code{HBEST_fast} is an MCMC algorithm that samples parameter values for HBEST (fast).
}
